fix: PR #280 — employmentTenureMinMonths type mismatch (float64 vs int)
Compile xətası:
  mygov_verify_service.go:293:15: invalid operation: months >= minMonths
  (mismatched types float64 and int)

Səbəb: PR #279-də minMonths parametri int tipindədir, amma months
float64-dır (time.Since().Hours() / 24 / 30).

Fix: float64(minMonths) ilə cast et.

Dəyişiklik (1 fayl, +1/-1):
- mygov_verify_service.go:293: months >= minMonths → months >= float64(minMonths)

+++ source/internal/service/mygov_verify_service.go
@@ -290,7 +290,7 @@ func checkEmploymentTenureFromEmployeeInfo(info *mygov.EmployeeInfoResponse, min
 	// Staj: başlama tarixindən bu günə, 30 günlük aylarla ("kesim" həddi 6 ay) — PR #255
 	months := time.Since(signDate).Hours() / 24 / 30
 
-	if months >= minMonths {
+	if months >= float64(minMonths) {
 		return true, months, fmt.Sprintf("İş yerində staj %.1f ay (%s, %s — ≥ %d ay) — uyğundur",
 			months, dateKind, employerName, minMonths)
 	}
